This is my very first project done without any help from anyone.very daunting and took me a while but it was a journey i cherish.
Ok so basically, this is a todo app that lets you schedule tasks to be done in the day, thus encouraging effectiveness and limits procrastination. You will be able to create accounts and have you personal todolist. I got templates from the web and designed the backend while making a lot of changes to the frontend too. If there is any area you feel i should improve on or any functionalities i should add, let me know. I wanna learn. This is the first of many.THANK YOU!!!!